+++ www/resources/js/app.js
@@ -20,7 +20,7 @@ window.Vue = require('vue');
 // const files = require.context('./', true, /\.vue$/i)
 // files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
 
-Vue.component('example-component', require('./components/ExampleComponent.vue').default);
+Vue.component('eyecatch-component', require('./components/EyecatchComponent.vue').default);
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
