+++ www/resources/js/components/EyecatchComponent.vue
